Fix Installer generated by HCC (BUG)
In previous versions of Hotcakes and there was a library called Hotcakes.Shipping.Ups.Keys, this was removed in the latest version but they forgot to remove the reference within the manifest so when we try to install the extension we get this error. I have corrected this within the manifest. I compiled the installer again and it installed successfully.

The solution was simply to delete the Assembly reference within the manifest.